feat(dapp): update tests
