+++ dapp/tests/helpers/helpers.ts
@@ -1,6 +1,7 @@
 // Copyright (c) 2025 IOTA Stiftung
 // SPDX-License-Identifier: Apache-2.0
 
+import { normalizeIotaName } from '@iota/iota-names-sdk';
 import { BrowserContext, Page } from '@playwright/test';
 
 import { expect } from '../helpers/fixtures';
@@ -36,4 +37,31 @@ export function purchaseNameFactory(): () => string {
         usedNames.add(name);
         return name;
     };
+}
+
+export async function purchaseName(page: Page, context: BrowserContext) {
+    const getPurchaseName = purchaseNameFactory();
+    const nameToPurchase = getPurchaseName();
+
+    await page.getByPlaceholder('Search for your IOTA name').click();
+
+    await buyNameFromAvailabilityCheckDialog(page, context, nameToPurchase);
+
+    const normalizedName = normalizeIotaName(nameToPurchase + '.iota', 'at', {
+        truncateLongParts: true,
+    });
+    await expect(page.getByText(`Successfully registered name ${normalizedName}`)).toBeVisible({
+        timeout: 60_000,
+    });
+    const closeIconDialog = page.getByTestId('close-icon');
+    if (await closeIconDialog.isVisible().catch(() => false)) {
+        await closeIconDialog.click();
+    }
+    const closeIcon = page.getByTestId('close-icon');
+    if (await closeIcon.isVisible().catch(() => false)) {
+        await closeIcon.click();
+    }
+    await page.keyboard.press('Escape').catch(() => {});
+    await page.waitForTimeout(200);
+    await page.goto('/my-names', { waitUntil: 'commit' });
 }
