+++ dapp/tests/management/management.spec.ts
@@ -1,13 +1,11 @@
 // Copyright (c) 2025 IOTA Stiftung
 // SPDX-License-Identifier: Apache-2.0
 
-import { normalizeIotaName } from '@iota/iota-names-sdk';
 import { formatAddress } from '@iota/iota-sdk/utils';
 import { expect } from '@playwright/test';
-import type { BrowserContext, Page } from '@playwright/test';
 
 import { test } from '../helpers/fixtures';
-import { buyNameFromAvailabilityCheckDialog, purchaseNameFactory } from '../helpers/helpers';
+import { purchaseName } from '../helpers/helpers';
 import {
     connectWallet,
     createWallet,
@@ -17,8 +15,6 @@ import {
 
 test.describe('Management Flow', () => {
     test.setTimeout(60_000);
-    const getPurchaseName = purchaseNameFactory();
-
     test.beforeAll(async ({ appPage, context, extensionPage, extensionName, sharedState }) => {
         await appPage.goto('/', { waitUntil: 'commit' });
 
@@ -35,45 +31,10 @@ test.describe('Management Flow', () => {
 
         sharedState.wallet.address = address;
         sharedState.wallet.mnemonic = mnemonic;
-
-        await appPage.close();
     });
-
-    test('Purchase name from Home', async ({ appPage: page, context }) => {
-        const nameToPurchase = getPurchaseName();
-
-        await page.getByPlaceholder('Search for your IOTA name').click();
-
-        await buyNameFromAvailabilityCheckDialog(page, context, nameToPurchase);
-
-        const normalizedName = normalizeIotaName(nameToPurchase + '.iota', 'at', {
-            truncateLongParts: true,
-        });
-        await expect(page.getByText(`Successfully registered name ${normalizedName}`)).toBeVisible({
-            timeout: 30_000,
-        });
-
-        await page.close();
-    });
-
-    // Helper to ensure there's at least one owned name to manage
-    async function ensureOwnedName(page: Page, context: BrowserContext) {
-        await page.getByRole('link', { name: 'My Names' }).click();
-        const count = await page.getByTestId('name-card').count();
-        if (count === 0) {
-            const nameToPurchase = getPurchaseName();
-            await page.getByPlaceholder('Search for your IOTA name').click();
-            await buyNameFromAvailabilityCheckDialog(page, context, nameToPurchase);
-            // Close purchase dialog if still open
-            const closeIcon = page.getByTestId('close-icon');
-            if (await closeIcon.isVisible().catch(() => false)) {
-                await closeIcon.click();
-            }
-        }
-    }
 
     test('Add a subname', async ({ appPage: page, context }) => {
-        await ensureOwnedName(page, context);
+        await purchaseName(page, context);
 
         const nameCard = page.getByTestId('name-card').first();
         await nameCard.getByTestId('name-card-avatar').hover();
@@ -101,7 +62,7 @@ test.describe('Management Flow', () => {
     });
 
     test('Connect address', async ({ appPage: page, context, sharedState }) => {
-        await ensureOwnedName(page, context);
+        await purchaseName(page, context);
 
         const nameCard = page.getByTestId('name-card').first();
         await nameCard.getByTestId('name-card-avatar').hover();
@@ -130,7 +91,7 @@ test.describe('Management Flow', () => {
     });
 
     test('Add an avatar', async ({ appPage: page, context }) => {
-        await ensureOwnedName(page, context);
+        await purchaseName(page, context);
 
         const nameCard = page.getByTestId('name-card').first();
         await nameCard.getByTestId('name-card-avatar').hover();
@@ -168,7 +129,7 @@ test.describe('Management Flow', () => {
     });
 
     test('Edit metadata', async ({ appPage: page, context }) => {
-        await ensureOwnedName(page, context);
+        await purchaseName(page, context);
 
         const nameCard = page.getByTestId('name-card').first();
         await nameCard.getByTestId('name-card-avatar').hover();
@@ -196,7 +157,7 @@ test.describe('Management Flow', () => {
     });
 
     test('View info', async ({ appPage: page, context }) => {
-        await ensureOwnedName(page, context);
+        await purchaseName(page, context);
 
         const nameCard = page.getByTestId('name-card').first();
         await nameCard.getByTestId('name-card-avatar').hover();
@@ -206,9 +167,10 @@ test.describe('Management Flow', () => {
 
         await page.getByText('View All Info', { exact: true }).click();
         const dialog = page.getByRole('dialog');
-        await expect(dialog.getByText('General Info')).toBeVisible();
+        await expect(dialog.getByText('All Info')).toBeVisible();
 
         await expect(dialog.getByText('Owner', { exact: false })).toBeVisible();
+        await expect(dialog.getByText('Object ID', { exact: false })).toBeVisible();
         await expect(dialog.getByText('Expiration Time', { exact: false })).toBeVisible();
 
         // Close dialog
